requirements.txt
google_search_api
PIL